[Refactor]#2 Constant 파일들을 data.constant 패키지로 이동
